move BlockUntilFork
