+++ common/forks/forks_test.go
@@ -24,3 +24,43 @@ func TestCollectSortedForkBlocks(t *testing.T) {
 		4: true,
 	}, m)
 }
+
+func TestBlocksUntilFork(t *testing.T) {
+	tests := map[string]struct {
+		block    uint64
+		forks    []uint64
+		expected uint64
+	}{
+		"NoFork": {
+			block:    44,
+			forks:    []uint64{},
+			expected: 0,
+		},
+		"BeforeFork": {
+			block:    0,
+			forks:    []uint64{1, 5},
+			expected: 1,
+		},
+		"OnFork": {
+			block:    1,
+			forks:    []uint64{1, 5},
+			expected: 4,
+		},
+		"OnLastFork": {
+			block:    5,
+			forks:    []uint64{1, 5},
+			expected: 0,
+		},
+		"AfterFork": {
+			block:    5,
+			forks:    []uint64{1, 5},
+			expected: 0,
+		},
+	}
+
+	for name, test := range tests {
+		t.Run(name, func(t *testing.T) {
+			require.Equal(t, test.expected, BlocksUntilFork(test.block, test.forks))
+		})
+	}
+}
