+++ rollup/internal/controller/watcher/chunk_proposer_test.go
@@ -7,7 +7,6 @@ import (
 
 	"github.com/scroll-tech/go-ethereum/params"
 	"github.com/stretchr/testify/assert"
-	"github.com/stretchr/testify/require"
 
 	"scroll-tech/common/database"
 	"scroll-tech/common/types"
@@ -195,46 +194,6 @@ func testChunkProposerLimits(t *testing.T) {
 					assert.Equal(t, chunks[0].Hash, block.ChunkHash)
 				}
 			}
-		})
-	}
-}
-
-func TestBlocksUntilFork(t *testing.T) {
-	tests := map[string]struct {
-		block    uint64
-		forks    []uint64
-		expected uint64
-	}{
-		"NoFork": {
-			block:    44,
-			forks:    []uint64{},
-			expected: 0,
-		},
-		"BeforeFork": {
-			block:    0,
-			forks:    []uint64{1, 5},
-			expected: 1,
-		},
-		"OnFork": {
-			block:    1,
-			forks:    []uint64{1, 5},
-			expected: 4,
-		},
-		"OnLastFork": {
-			block:    5,
-			forks:    []uint64{1, 5},
-			expected: 0,
-		},
-		"AfterFork": {
-			block:    5,
-			forks:    []uint64{1, 5},
-			expected: 0,
-		},
-	}
-
-	for name, test := range tests {
-		t.Run(name, func(t *testing.T) {
-			require.Equal(t, test.expected, blocksUntilFork(test.block, test.forks))
 		})
 	}
 }
